BUG: Make DeformableModel Compiled
BUG: Fix whitespace and remove redundant commands in CMakeLists.txt

+++ DeformableModelSimplexMesh/ImageSurfaceViewer.h
@@ -18,7 +18,7 @@
 
 #ifndef vtkFloatingPointType
 # define vtkFloatingPointType vtkFloatingPointType
-typedef float vtkFloatingPointType;
+typedef double vtkFloatingPointType;
 #endif
 
   /**
